loss function fix
Previously, the magnitude of the loss function changed a lot with the
training data, and empirically that seemed to make it hard to set the
learning rate in way that was robust to changing the size and content of
the training set. This commit scales components of the loss in a way
that makes it more predictable to alleviate that problem.