expand practical range of auxiliary predictions
Previously, we went to some effort to ensure that the calculation of the
auxiliary predictions had both exponential discounting and a hard upper
limit of 1. Practically speaking, since we want a relatively low
discounting rate (because the exact timing of events is not super fixed
in this game), this meant that the scaling required left us with most
non-zero values being very close to 0. Since most values are zero in the
first place, that doesn't give the net much signal to work with.

So this is an attempt to boost the signal. We keep the low discounting
rate, but don't bother scaling to ensure that sums stay under 1 --
instead, if the values go over 1, we just clip. This seems like a pretty
good compromise in practice; we expect clipping to happen rarely, and
when it does happen it's because we really want the net to confidently
predict that event will happen anyway.