simplify preprocessing logic
The previous preprocessing logic was kinda twisted. Working out why it
was correct was a mental exercise every time it came up, and it led to
several confusing reports in the GUI and in saved games. This new
preprocessor is conceptually much simpler: do something special the
first time you visit a node and that's it.

The special thing for RNG nodes is to do pathfinding and replicate the
result across all the possible RNG choices. The special thing for pill
placement nodes is to run the net to evaluate the position and record
the prior probabilities of each choice. Bonus: we've eliminated the
unsafeInterleaveIO that used to be associated with scheduling net
evaluations in the preprocessor.

One side effect of this simplification is that each step of the search
requires only one evaluation of the net rather than nine like the
previous version. Consequently we can bump up the number of iterations
per move significantly; we also bump up the number of generation threads
significantly because each thread puts 9x less work on the queue before
blocking.