revert "train on everything until everything is too big"
There were just too many accounting annoyances for this to be worth it.
You gotta report the actual number of tensors loaded, then increase the
SearchSpeed by the right amount; and then there's getting the infrequent
events to still happen infrequently which is even more of a nightmare
when you're starting out and the number of tensors per iteration is
small.